+++ lib/rubocop/cop/style/while_until_modifier.rb
@@ -5,6 +5,7 @@ module Rubocop
     module Style
       # Checks for while and until statements that would fit on one line
       # if written as a modifier while/until.
+      # The maximum line length is configurable.
       class WhileUntilModifier < Cop
         include StatementModifier